renamed to core/errorHandler.js
