+++ src/amo/components/OverallRating.js
@@ -4,7 +4,7 @@ import { compose } from 'redux';
 import { withRouter } from 'react-router';
 import classNames from 'classnames';
 
-import { withErrorHandling } from 'core/errors';
+import { withErrorHandling } from 'core/errorHandler';
 import { setReview } from 'amo/actions/reviews';
 import { getLatestUserReview, submitReview } from 'amo/api';
 import translate from 'core/i18n/translate';
